refactor(common) move 'registry' events module in 'events' module
Group events.
The distinction between bus-servers and registry isn't as practical.
