+++ atspi-common/src/events/registry.rs
