+++ atspi-common/src/lib.rs
@@ -29,9 +29,6 @@ pub mod operation;
 pub use operation::Operation;
 pub mod interface;
 pub use interface::{Interface, InterfaceSet};
-pub mod registry;
-pub use registry::socket::AvailableEvent;
-pub use registry::{EventListenerDeregisteredEvent, EventListenerRegisteredEvent};
 pub mod state;
 pub use state::{State, StateSet};
 pub mod cache;
